requirements.txt
albumentations==1.1.0
backports.entry-points-selectable==1.1.0
CacheControl==0.12.9
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.7
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cycler==0.11.0
distlib==0.3.3
filelock==3.3.2
html5lib==1.1
idna==3.3
imageio==2.10.3
-e git+https://github.com/anvelezec/tf-challenge-iris.git@bb75465cca55a1155007c3a6d76749b5102b687e#egg=Iris
joblib==1.1.0
keyring==21.8.0
kiwisolver==1.3.2
lockfile==0.12.2
matplotlib==3.4.3
msgpack==1.0.2
networkx==2.6.3
numpy==1.21.4
opencv-python-headless==4.5.4.58
packaging==20.9
pandas==1.3.4
pastel==0.2.1
pexpect==4.8.0
Pillow==8.4.0
pkginfo==1.7.1
platformdirs==2.4.0
poetry==1.1.11
poetry-core==1.0.7
ptyprocess==0.7.0
pylev==1.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-toolbelt==0.9.1
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
shellingham==1.4.0
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomlkit==0.7.2
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
